+++ src/main/java/com/example/communityapplication/service/ContentTemplateService.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface ContentTemplateService {
     public List<ContentTemplate> getByCommunity(Community community);
-    ContentTemplate getById(Long id);
+    ContentTemplate getById(int id);
 }
